fix: searchParams prerender issue
