+++ src/app/(before-login)/welcome/_component/content.module.scss
@@ -1,7 +1,7 @@
 @use "@/styles/color.scss" as *;
 @use "@/styles/font.scss" as *;
 
-.welcome {
+.content {
   &__container {
     display: flex;
     justify-content: center;
