+++ src/app/course-review/page.tsx
@@ -2,34 +2,34 @@ import Tabs from '@/components/common/taps';
 import Layout from '@/components/layout';
 import ReviewLayout from '@/components/review-layout';
 
-import {
-  dummyModalReviewsDataUnWrite,
-  dummyModalReviewsDataWrite,
-} from './dummy';
+// import {
+//   dummyModalReviewsDataUnWrite,
+//   dummyModalReviewsDataWrite,
+// } from './dummy';
 
 const Review = () => {
   const tapsFirstTitle = `Unwritten Review()`;
   const tapsSecondTitle = `written Review()`;
-  const tabsItem = [
-    {
-      title: tapsFirstTitle,
-      content: (
-        <ReviewLayout
-          modal='unWrite'
-          modalReviewsData={dummyModalReviewsDataUnWrite}
-        />
-      ),
-    },
-    {
-      title: tapsSecondTitle,
-      content: (
-        <ReviewLayout
-          modal='write'
-          modalReviewsData={dummyModalReviewsDataWrite}
-        />
-      ),
-    },
-  ];
+  // const tabsItem = [
+  //   {
+  //     title: tapsFirstTitle,
+  //     content: (
+  //       <ReviewLayout
+  //         modal='unWrite'
+  //         modalReviewsData={dummyModalReviewsDataUnWrite}
+  //       />
+  //     ),
+  //   },
+  //   {
+  //     title: tapsSecondTitle,
+  //     content: (
+  //       <ReviewLayout
+  //         modal='write'
+  //         modalReviewsData={dummyModalReviewsDataWrite}
+  //       />
+  //     ),
+  //   },
+  // ];
   return (
     <Layout
       hasTopNav={true}
@@ -37,7 +37,8 @@ const Review = () => {
       back={true}
       title='My course review'
     >
-      <Tabs tabs={tabsItem} version={1} />
+      <div></div>
+      {/* <Tabs tabs={tabsItem} version={1} /> */}
     </Layout>
   );
 };
